Began working on an approach to lint as you type
Using neomake with neovim (which support asynchronous jobs), there is
now a hidden buffer which is written to a temporary file as you type
such that the buffer can be saved and linted. This is needed because
linters generally cannot run on the buffer in memory and need a file on
disk.

There is likely still more work to do, but it's a nice first step.